refactor(agents): extract FileMaterializer and AgentEmailSender from chat-sdk
