+++ apps/api/src/app/agents/agents.module.ts
@@ -28,7 +28,9 @@ import { AgentsController } from './agents.controller';
 import { AgentsMcpOAuthController } from './agents-mcp-oauth.controller';
 import { AgentsPublicController } from './agents-public.controller';
 import { AgentsWebhookController } from './agents-webhook.controller';
+import { FileMaterializer } from './conversation-runtime/egress/file-materializer.service';
 import { OutboundGateway } from './conversation-runtime/egress/outbound.gateway';
+import { AgentEmailSender } from './email/agent-email-sender.service';
 import { AgentRuntimeExceptionFilter } from './filters/agent-runtime-exception.filter';
 import { AgentAttachmentStorage } from './services/agent-attachment-storage.service';
 import { AgentConfigResolver } from './services/agent-config-resolver.service';
@@ -90,6 +92,8 @@ import { USE_CASES } from './usecases';
     McpConnectionVaultService,
     DemoClaudeQuotaPolicy,
     ChatSdkService,
+    FileMaterializer,
+    AgentEmailSender,
     OutboundGateway,
     McpOAuthDiscoveryService,
     TelegramMobileLinkTokenService,
