+++ apps/api/src/app/agents/services/chat-sdk.service.ts
@@ -1,28 +1,13 @@
-import { randomUUID } from 'node:crypto';
-import * as dns from 'node:dns';
-import * as http from 'node:http';
-import * as https from 'node:https';
 import { BadGatewayException, BadRequestException, Injectable, OnModuleDestroy } from '@nestjs/common';
-import {
-  areNovuEmailCredentialsSet,
-  assertSafeOutboundUrl,
-  buildAgentSharedInbox,
-  CacheService,
-  CalculateLimitNovuIntegration,
-  decryptCredentials,
-  isAgentSharedInboxEnabled,
-  isPrivateIp,
-  MailFactory,
-  PinoLogger,
-  SsrfBlockedError,
-} from '@novu/application-generic';
-import { IntegrationEntity, IntegrationRepository, MessageRepository } from '@novu/dal';
+import { CacheService, PinoLogger } from '@novu/application-generic';
 import type { AgentAction, SentMessageInfo } from '@novu/framework';
-import { ChannelTypeEnum, EmailProviderIdEnum, type IEmailOptions } from '@novu/shared';
 import type { AdapterPostableMessage, Chat, EmojiValue, Message, PlanModel, ReactionEvent, Thread } from 'chat';
 import { Request as ExpressRequest, Response as ExpressResponse } from 'express';
 import { LRUCache } from 'lru-cache';
-import type { FileRef, ReplyContentDto } from '../shared/dtos/agent-reply-payload.dto';
+import type { ChatSdkFile, ChatSdkReplyContent } from '../conversation-runtime/egress/file-materializer.service';
+import { FileMaterializer } from '../conversation-runtime/egress/file-materializer.service';
+import { AgentEmailSender, resolveAgentEmailSenderName } from '../email/agent-email-sender.service';
+import type { ReplyContentDto } from '../shared/dtos/agent-reply-payload.dto';
 import { AgentPlatformEnum } from '../shared/enums/agent-platform.enum';
 import { captureAgentException, captureAgentWarning } from '../shared/errors/capture-agent-sentry';
 import { esmImport } from '../shared/util/esm-import';
@@ -73,17 +58,6 @@ function toDeliveryError(err: unknown): never {
     error: 'delivery_failed',
     message: detail ? `${base}: ${detail}` : base,
   });
-}
-
-/** Ensure a Message-ID value is wrapped in RFC 5322 angle brackets. */
-function wrapMsgId(id: string): string {
-  const trimmed = id.trim();
-
-  return trimmed.startsWith('<') && trimmed.endsWith('>') ? trimmed : `<${trimmed}>`;
-}
-
-function resolveAgentEmailSenderName(config: ResolvedAgentConfig): string {
-  return config.credentials.senderName?.trim() || config.agentName;
 }
 
 /**
@@ -138,30 +112,6 @@ function extractRecipientFromThreadId(threadId: string): string {
 
 const MAX_CACHED_INSTANCES = 200;
 const INSTANCE_TTL_MS = 1000 * 60 * 30;
-const BASE64_REGEX = /^[A-Za-z0-9+/]*={0,2}$/;
-const MAX_INLINE_FILE_BYTES = 5 * 1024 * 1024;
-const MAX_INLINE_AGGREGATE_FILE_BYTES = 5 * 1024 * 1024;
-const MAX_FILE_BYTES = 25 * 1024 * 1024;
-const MAX_FILES_PER_MESSAGE = 15;
-const MAX_AGGREGATE_FILE_BYTES = 50 * 1024 * 1024;
-const MAX_INLINE_FILE_BASE64_CHARS = 7_000_000;
-const FILE_FETCH_TIMEOUT_MS = 10_000;
-const MAX_FILE_FETCH_REDIRECTS = 3;
-const SUPPORTED_FILE_PLATFORMS = new Set<string>([
-  AgentPlatformEnum.SLACK,
-  AgentPlatformEnum.TEAMS,
-  AgentPlatformEnum.WHATSAPP,
-]);
-const UNSUPPORTED_FILE_PLATFORMS = new Set<string>([AgentPlatformEnum.EMAIL]);
-// EMAIL_ALTERNATIVES_SUPPORTED_PROVIDERS is a deliberate allowlist for providers that preserve custom MIME
-// alternatives used by Gmail reactions; Braze, Brevo, Mailgun, Mailjet, Mailtrap, Mandrill, Plunk, Postmark,
-// Resend, SparkPost, and similar providers are excluded until their SDK paths are verified.
-const EMAIL_ALTERNATIVES_SUPPORTED_PROVIDERS = new Set<string>([
-  EmailProviderIdEnum.CustomSMTP,
-  EmailProviderIdEnum.Outlook365,
-  EmailProviderIdEnum.SendGrid,
-  EmailProviderIdEnum.SES,
-]);
 
 /**
  * Holds a cached Chat instance alongside a mutable pointer to the current
@@ -181,16 +131,6 @@ interface CachedChat {
   adapterFingerprint: string;
 }
 
-type ChatSdkFile = Omit<FileRef, 'data'> & { data?: Buffer };
-type ChatSdkReplyContent = Omit<ReplyContentDto, 'files'> & { files?: ChatSdkFile[] };
-type MaterializedFile = ChatSdkFile & { size: number; source: 'data' | 'url' };
-type PinnedFileResponse = {
-  status: number;
-  statusText: string;
-  headers: http.IncomingHttpHeaders;
-  data: Buffer;
-};
-
 @Injectable()
 export class ChatSdkService implements OnModuleDestroy {
   private readonly instances: LRUCache<string, CachedChat>;
@@ -201,10 +141,9 @@ export class ChatSdkService implements OnModuleDestroy {
     private readonly logger: PinoLogger,
     private readonly cacheService: CacheService,
     private readonly agentConfigResolver: AgentConfigResolver,
-    private readonly integrationRepository: IntegrationRepository,
     private readonly actionTokenService: AgentEmailActionTokenService,
-    private readonly calculateLimitNovuIntegration: CalculateLimitNovuIntegration,
-    private readonly messageRepository: MessageRepository
+    private readonly fileMaterializer: FileMaterializer,
+    private readonly agentEmailSender: AgentEmailSender
   ) {
     this.logger.setContext(this.constructor.name);
     this.instances = new LRUCache<string, CachedChat>({
@@ -343,7 +282,7 @@ export class ChatSdkService implements OnModuleDestroy {
     // avoid rehydrating from a serialized blob and don't trip the "No Chat singleton
     // registered" check that `ThreadImpl.fromJSON` hits for card/postable replies.
     const thread = chat.thread(platformThreadId);
-    const deliveryContent = await this.prepareContentForDelivery(content, platform, agentId);
+    const deliveryContent = await this.fileMaterializer.prepareContentForDelivery(content, platform, agentId);
 
     const postArg = this.buildAdapterPostableMessage(deliveryContent);
 
@@ -381,7 +320,7 @@ export class ChatSdkService implements OnModuleDestroy {
     const chat = await this.getOrCreate(instanceKey, agentId, config.platform, config);
 
     const dmThread = await chat.openDM(platformUserId);
-    const deliveryContent = await this.prepareContentForDelivery(content, config.platform, agentId);
+    const deliveryContent = await this.fileMaterializer.prepareContentForDelivery(content, config.platform, agentId);
 
     const postArg = this.buildAdapterPostableMessage(deliveryContent);
 
@@ -412,7 +351,7 @@ export class ChatSdkService implements OnModuleDestroy {
       throw new BadRequestException(`Platform ${platform} does not support editing messages`);
     }
 
-    const deliveryContent = await this.prepareContentForDelivery(content, platform, agentId);
+    const deliveryContent = await this.fileMaterializer.prepareContentForDelivery(content, platform, agentId);
 
     const editPayload = this.buildAdapterPostableMessage(deliveryContent);
 
@@ -473,79 +412,6 @@ export class ChatSdkService implements OnModuleDestroy {
     await adapter.editObject(platformThreadId, platformMessageId, 'plan', model).catch(toDeliveryError);
   }
 
-  private async prepareContentForDelivery(
-    content: ReplyContentDto,
-    platform: string = AgentPlatformEnum.SLACK,
-    agentId?: string
-  ): Promise<ChatSdkReplyContent> {
-    if (!content.files?.length) {
-      return content as ChatSdkReplyContent;
-    }
-
-    if (UNSUPPORTED_FILE_PLATFORMS.has(platform)) {
-      this.logger.warn(
-        {
-          agentId,
-          platform,
-          droppedCount: content.files.length,
-        },
-        'Dropping outbound agent files because platform does not support attachments'
-      );
-
-      const { files: _files, ...withoutFiles } = content;
-
-      return withoutFiles as ChatSdkReplyContent;
-    }
-
-    if (!SUPPORTED_FILE_PLATFORMS.has(platform)) {
-      throw new BadRequestException({
-        error: 'attachment_failed',
-        message: `File attachments are not supported on platform "${platform}".`,
-      });
-    }
-
-    if (content.files.length > MAX_FILES_PER_MESSAGE) {
-      throw new BadRequestException({
-        error: 'attachment_failed',
-        message: `Too many attachments: maximum is ${MAX_FILES_PER_MESSAGE} files per message.`,
-      });
-    }
-
-    const files: ChatSdkFile[] = [];
-    let aggregateSize = 0;
-    let inlineAggregateSize = 0;
-
-    for (const [index, file] of content.files.entries()) {
-      const materialized = await this.prepareFileForDelivery(file, index);
-      aggregateSize += materialized.size;
-      if (materialized.source === 'data') {
-        inlineAggregateSize += materialized.size;
-      }
-
-      if (aggregateSize > MAX_AGGREGATE_FILE_BYTES) {
-        throw new BadRequestException({
-          error: 'attachment_failed',
-          message: `Total attachment size exceeds ${this.formatBytes(MAX_AGGREGATE_FILE_BYTES)}.`,
-        });
-      }
-
-      if (inlineAggregateSize > MAX_INLINE_AGGREGATE_FILE_BYTES) {
-        throw new BadRequestException({
-          error: 'attachment_failed',
-          message: `Total inline attachment size exceeds ${this.formatBytes(MAX_INLINE_AGGREGATE_FILE_BYTES)}. Use URLs for larger files.`,
-        });
-      }
-
-      const { size: _size, source: _source, ...chatSdkFile } = materialized;
-      files.push(chatSdkFile);
-    }
-
-    return {
-      ...content,
-      files,
-    };
-  }
-
   private buildAdapterPostableMessage(deliveryContent: ChatSdkReplyContent): AdapterPostableMessage {
     if (deliveryContent.card) {
       const payload: { card: unknown; files?: ChatSdkFile[] } = {
@@ -563,294 +429,6 @@ export class ChatSdkService implements OnModuleDestroy {
       markdown: deliveryContent.markdown ?? '',
       files: deliveryContent.files,
     } as unknown as AdapterPostableMessage;
-  }
-
-  private async prepareFileForDelivery(file: FileRef, index: number): Promise<MaterializedFile> {
-    const data = (file as { data?: unknown }).data;
-    const url = (file as { url?: unknown }).url;
-
-    if (data !== undefined && data !== null) {
-      if (typeof data !== 'string') {
-        throw new BadRequestException({
-          error: 'attachment_failed',
-          message: `Invalid file ${this.describeFile(file, index)}: data must be a base64-encoded string.`,
-        });
-      }
-
-      const buffer = this.decodeBase64FileData(data, file, index);
-      const { url: _url, ...fileWithoutUrl } = file;
-
-      return {
-        ...fileWithoutUrl,
-        data: buffer,
-        size: buffer.length,
-        source: 'data',
-      };
-    }
-
-    if (typeof url !== 'string') {
-      throw new BadRequestException({
-        error: 'attachment_failed',
-        message: `Invalid file ${this.describeFile(file, index)}: provide a public HTTP(S) url or base64 data.`,
-      });
-    }
-
-    const fetched = await this.fetchFileUrl(url, file, index);
-    const { url: _url, ...fileWithoutUrl } = file;
-
-    return {
-      ...fileWithoutUrl,
-      data: fetched.data,
-      mimeType: file.mimeType || fetched.mimeType,
-      size: fetched.data.length,
-      source: 'url',
-    };
-  }
-
-  private decodeBase64FileData(data: string, file: FileRef, index: number): Buffer {
-    const normalized = data.replace(/\s/g, '');
-    const remainder = normalized.length % 4;
-
-    if (normalized.length > MAX_INLINE_FILE_BASE64_CHARS) {
-      throw new BadRequestException({
-        error: 'attachment_failed',
-        message: `Invalid file ${this.describeFile(file, index)}: inline data must be ${this.formatBytes(MAX_INLINE_FILE_BYTES)} or smaller.`,
-      });
-    }
-
-    if (!normalized || remainder === 1 || !BASE64_REGEX.test(normalized)) {
-      throw new BadRequestException({
-        error: 'attachment_failed',
-        message: `Invalid file ${this.describeFile(file, index)}: data must be a base64-encoded string.`,
-      });
-    }
-
-    const padded = remainder === 0 ? normalized : normalized.padEnd(normalized.length + (4 - remainder), '=');
-    const buffer = Buffer.from(padded, 'base64');
-
-    if (buffer.toString('base64').replace(/=+$/, '') !== normalized.replace(/=+$/, '')) {
-      throw new BadRequestException({
-        error: 'attachment_failed',
-        message: `Invalid file ${this.describeFile(file, index)}: data must be a base64-encoded string.`,
-      });
-    }
-
-    if (buffer.length > MAX_INLINE_FILE_BYTES) {
-      throw new BadRequestException({
-        error: 'attachment_failed',
-        message: `Invalid file ${this.describeFile(file, index)}: inline data must be ${this.formatBytes(MAX_INLINE_FILE_BYTES)} or smaller.`,
-      });
-    }
-
-    return buffer;
-  }
-
-  private async fetchFileUrl(url: string, file: FileRef, index: number): Promise<{ data: Buffer; mimeType?: string }> {
-    const response = await this.fetchValidatedFileUrl(url, file, index);
-
-    if (response.status < 200 || response.status >= 300) {
-      throw new BadRequestException({
-        error: 'attachment_failed',
-        message: `Failed to fetch file ${this.describeFile(file, index)}: ${response.status} ${response.statusText}`,
-      });
-    }
-
-    const contentLength = this.getHeader(response.headers, 'content-length');
-    if (contentLength) {
-      const size = Number(contentLength);
-      if (Number.isFinite(size) && size > MAX_FILE_BYTES) {
-        throw new BadRequestException({
-          error: 'attachment_failed',
-          message: `Invalid file ${this.describeFile(file, index)}: file size exceeds ${this.formatBytes(MAX_FILE_BYTES)}.`,
-        });
-      }
-    }
-
-    const data = response.data;
-    const mimeType = this.getHeader(response.headers, 'content-type');
-
-    return { data, mimeType };
-  }
-
-  private async fetchValidatedFileUrl(url: string, file: FileRef, index: number): Promise<PinnedFileResponse> {
-    let currentUrl = url;
-
-    for (let redirectCount = 0; redirectCount <= MAX_FILE_FETCH_REDIRECTS; redirectCount += 1) {
-      const ssrfError = await this.validateFileUrl(currentUrl);
-      if (ssrfError) {
-        throw new BadRequestException({
-          error: 'attachment_failed',
-          message: `Invalid file ${this.describeFile(file, index)} url: ${ssrfError}`,
-        });
-      }
-
-      let response: PinnedFileResponse;
-      try {
-        response = await this.requestPinnedFileUrl(currentUrl, file, index);
-      } catch (err) {
-        if (err instanceof BadRequestException) {
-          throw err;
-        }
-
-        const message = err instanceof Error ? err.message : String(err);
-        throw new BadRequestException({
-          error: 'attachment_failed',
-          message: `Failed to fetch file ${this.describeFile(file, index)}: ${message}`,
-        });
-      }
-
-      if (response.status < 300 || response.status >= 400) {
-        return response;
-      }
-
-      const location = this.getHeader(response.headers, 'location');
-      if (!location) {
-        throw new BadRequestException({
-          error: 'attachment_failed',
-          message: `Failed to fetch file ${this.describeFile(file, index)}: redirect response missing Location header.`,
-        });
-      }
-
-      currentUrl = new URL(location, currentUrl).toString();
-    }
-
-    throw new BadRequestException({
-      error: 'attachment_failed',
-      message: `Failed to fetch file ${this.describeFile(file, index)}: too many redirects.`,
-    });
-  }
-
-  private async validateFileUrl(url: string): Promise<string | null> {
-    try {
-      assertSafeOutboundUrl(url);
-    } catch (err) {
-      if (err instanceof SsrfBlockedError) {
-        return err.message;
-      }
-      throw err;
-    }
-
-    return null;
-  }
-
-  private async requestPinnedFileUrl(url: string, file: FileRef, index: number): Promise<PinnedFileResponse> {
-    const parsed = new URL(url);
-    const address = await this.resolvePublicAddress(parsed, file, index);
-    const client = parsed.protocol === 'https:' ? https : http;
-
-    return await new Promise((resolve, reject) => {
-      const request = client.request(
-        {
-          protocol: parsed.protocol,
-          hostname: address.address,
-          family: address.family,
-          port: parsed.port || undefined,
-          path: `${parsed.pathname}${parsed.search}`,
-          method: 'GET',
-          headers: { Host: parsed.host },
-          servername: parsed.hostname,
-          timeout: FILE_FETCH_TIMEOUT_MS,
-        },
-        (response) => {
-          const status = response.statusCode ?? 0;
-          const statusText = response.statusMessage ?? '';
-
-          if (status >= 300 && status < 400) {
-            response.resume();
-            resolve({ status, statusText, headers: response.headers, data: Buffer.alloc(0) });
-
-            return;
-          }
-
-          const contentLength = this.getHeader(response.headers, 'content-length');
-          if (contentLength) {
-            const size = Number(contentLength);
-            if (Number.isFinite(size) && size > MAX_FILE_BYTES) {
-              response.destroy();
-              reject(
-                new BadRequestException({
-                  error: 'attachment_failed',
-                  message: `Invalid file ${this.describeFile(file, index)}: file size exceeds ${this.formatBytes(MAX_FILE_BYTES)}.`,
-                })
-              );
-
-              return;
-            }
-          }
-
-          const chunks: Buffer[] = [];
-          let total = 0;
-
-          response.on('data', (chunk: Buffer) => {
-            total += chunk.length;
-            if (total > MAX_FILE_BYTES) {
-              response.destroy(
-                new BadRequestException({
-                  error: 'attachment_failed',
-                  message: `Invalid file ${this.describeFile(file, index)}: file size exceeds ${this.formatBytes(MAX_FILE_BYTES)}.`,
-                })
-              );
-
-              return;
-            }
-
-            chunks.push(chunk);
-          });
-          response.on('end', () =>
-            resolve({ status, statusText, headers: response.headers, data: Buffer.concat(chunks, total) })
-          );
-          response.on('error', reject);
-        }
-      );
-
-      request.on('timeout', () => request.destroy(new Error('Request timed out')));
-      request.on('error', reject);
-      request.end();
-    });
-  }
-
-  private async resolvePublicAddress(parsed: URL, file: FileRef, index: number): Promise<dns.LookupAddress> {
-    let addresses: dns.LookupAddress[];
-    try {
-      addresses = await dns.promises.lookup(parsed.hostname, { all: true });
-    } catch {
-      throw new BadRequestException({
-        error: 'attachment_failed',
-        message: `Invalid file ${this.describeFile(file, index)} url: Unable to resolve hostname "${parsed.hostname}".`,
-      });
-    }
-
-    if (!addresses.length) {
-      throw new BadRequestException({
-        error: 'attachment_failed',
-        message: `Invalid file ${this.describeFile(file, index)} url: Unable to resolve hostname "${parsed.hostname}".`,
-      });
-    }
-
-    for (const { address } of addresses) {
-      if (isPrivateIp(address)) {
-        throw new BadRequestException({
-          error: 'attachment_failed',
-          message: `Invalid file ${this.describeFile(file, index)} url: Requests to private or reserved IP addresses are not allowed (resolved: ${address}).`,
-        });
-      }
-    }
-
-    return addresses[0];
-  }
-
-  private getHeader(headers: http.IncomingHttpHeaders, name: string): string | undefined {
-    const value = headers[name.toLowerCase()];
-
-    return Array.isArray(value) ? value[0] : value;
-  }
-
-  private describeFile(file: FileRef, index: number): string {
-    return file.filename ? `"${file.filename}"` : `at index ${index}`;
-  }
-
-  private formatBytes(bytes: number): string {
-    return `${Math.floor(bytes / (1024 * 1024))} MB`;
   }
 
   async removeReaction(
@@ -1003,300 +581,6 @@ export class ChatSdkService implements OnModuleDestroy {
     });
   }
 
-  private buildSendEmailCallback(
-    config: ResolvedAgentConfig,
-    outboundIntegrationId: string | undefined
-  ): (params: {
-    from: string;
-    to: string;
-    subject: string;
-    html: string;
-    text?: string;
-    alternatives?: Array<{
-      contentType: string;
-      content: string | Buffer;
-    }>;
-    inReplyTo?: string;
-    references?: string;
-    messageId?: string;
-  }) => Promise<{ messageId?: string }> {
-    return async (params) => {
-      if (!outboundIntegrationId) {
-        throw new BadRequestException(
-          'Email agent integration is missing outboundIntegrationId. Reconfigure the agent email setup.'
-        );
-      }
-
-      const integration = await this.integrationRepository.findOne({
-        _id: outboundIntegrationId,
-        _environmentId: config.environmentId,
-        _organizationId: config.organizationId,
-        channel: ChannelTypeEnum.EMAIL,
-      });
-
-      if (!integration) {
-        throw new BadRequestException(
-          `Outbound email integration ${outboundIntegrationId} not found or does not belong to this environment`
-        );
-      }
-
-      if (integration.providerId === EmailProviderIdEnum.NovuAgent) {
-        throw new BadRequestException(
-          `Integration ${outboundIntegrationId} is the inbound NovuAgent provider and cannot be used as an outbound sender`
-        );
-      }
-
-      if (!integration.active) {
-        throw new BadRequestException(
-          `Outbound email integration ${outboundIntegrationId} (${integration.providerId}) is inactive`
-        );
-      }
-
-      if (integration.providerId === EmailProviderIdEnum.Novu) {
-        return this.sendViaNovuDemoProvider(config, params, integration);
-      }
-
-      const hasUnsupportedAlternatives =
-        params.alternatives?.length && !EMAIL_ALTERNATIVES_SUPPORTED_PROVIDERS.has(integration.providerId);
-      if (hasUnsupportedAlternatives) {
-        // NovuEmailAdapterImpl.addReaction supplies a reaction Message-ID; any custom MIME alternative caller must do
-        // the same so skipped unsupported sends don't claim provider delivery.
-        if (!params.messageId) {
-          this.logger.warn(
-            {
-              providerId: integration.providerId,
-              outboundIntegrationId,
-            },
-            'Skipping email with custom MIME alternatives because the outbound provider is unsupported and no messageId was supplied'
-          );
-
-          return { messageId: undefined };
-        }
-
-        this.logger.warn(
-          {
-            providerId: integration.providerId,
-            outboundIntegrationId,
-          },
-          'Skipping email reaction because the outbound provider does not support custom MIME alternatives'
-        );
-
-        return { messageId: params.messageId };
-      }
-
-      const decrypted = decryptCredentials(integration.credentials);
-
-      // The chat-adapter-email contract guarantees params.from is the agent's inbound address
-      // (see packages/chat-adapter-email/src/adapter.ts postMessage/addReaction). We treat it as
-      // the Reply-To target so subscriber replies still reach the agent's inbox even when the
-      // outbound From is rewritten to the sending provider's configured sender (or a per-agent
-      // override). When neither override nor outbound.from is set, we fall back to the agent
-      // address for From and skip Reply-To — preserving the legacy behavior.
-      // Prefer the shared-inbox address (cloud) so Reply-To always routes back to the agent
-      // even when the SDK has no DomainRoute-derived address handy.
-      const agentInboundAddress = this.resolveAgentInboundAddress(config, params.from);
-      const overrideFrom = config.credentials.useFromAddressOverride
-        ? config.credentials.fromAddressOverride?.trim() || undefined
-        : undefined;
-      const outboundFrom = (decrypted.from as string | undefined)?.trim() || undefined;
-      const effectiveFrom = overrideFrom || agentInboundAddress || outboundFrom;
-      const replyToHeader = effectiveFrom !== agentInboundAddress ? agentInboundAddress : undefined;
-      const senderName = resolveAgentEmailSenderName(config);
-
-      const mailFactory = new MailFactory();
-      const handler = mailFactory.getHandler({ ...integration, credentials: decrypted }, effectiveFrom);
-
-      const mailOptions: IEmailOptions = {
-        to: [params.to],
-        subject: params.subject,
-        html: params.html,
-        text: params.text,
-        alternatives: params.alternatives,
-        from: effectiveFrom,
-        ...(replyToHeader ? { replyTo: replyToHeader } : {}),
-        senderName,
-        headers: {
-          ...(params.messageId ? { 'Message-ID': wrapMsgId(params.messageId) } : {}),
-          ...(params.inReplyTo ? { 'In-Reply-To': wrapMsgId(params.inReplyTo) } : {}),
-          ...(params.references
-            ? { References: params.references.split(/\s+/).filter(Boolean).map(wrapMsgId).join(' ') }
-            : {}),
-        },
-      };
-
-      const result = await handler.send(mailOptions).catch(toDeliveryError);
-
-      return { messageId: result?.id || params.messageId || '' };
-    };
-  }
-
-  /**
-   * Resolve the canonical inbound address used for Reply-To. Preference order:
-   *
-   *   1. The synthetic shared inbox `{slug}-{inboxRoutingKey}@<shared-domain>`
-   *      when the cloud feature is enabled and the shared inbox itself is not
-   *      disabled. System-managed and always works.
-   *   2. The fallback supplied by the chat-adapter-email SDK — already a
-   *      custom-domain agent route configured by the user (the SDK builds it
-   *      from `DomainRoute` rows), or whatever the platform passed on
-   *      self-hosted.
-   *
-   * Replies must always reach an inbox the worker will actually process, so
-   * we deliberately do not return the shared inbox here when
-   * `sharedInboxDisabled` is set — the worker would drop those messages and
-   * we fall through to the SDK's custom-domain address instead.
-   */
-  private resolveAgentInboundAddress(config: ResolvedAgentConfig, fallback: string): string {
-    const slug = config.credentials.emailSlugPrefix;
-    const inboxRoutingKey = config.credentials.inboxRoutingKey;
-    const sharedDisabled = Boolean(config.credentials.sharedInboxDisabled);
-    if (isAgentSharedInboxEnabled() && slug && inboxRoutingKey && !sharedDisabled) {
-      try {
-        return buildAgentSharedInbox(slug, inboxRoutingKey);
-      } catch (err) {
-        this.logger.warn({ err, agentId: config.agentId }, 'Falling back to params.from - shared inbox build failed');
-        captureAgentWarning(err, {
-          component: 'chat-sdk',
-          operation: 'resolve-agent-inbound-address',
-          agentId: config.agentId,
-        });
-      }
-    }
-
-    return fallback;
-  }
-
-  /**
-   * Outbound demo path: the agent is wired to the bundled Novu Email demo
-   * provider row. We override the integration's stored credentials with the
-   * cloud demo API key (`NOVU_EMAIL_INTEGRATION_API_KEY`) and force `From` to
-   * `{slug}-{inboxRoutingKey}@<shared-domain>` so replies route back to the
-   * same inbox. Quota-gated by the same per-environment 300/month cap as
-   * workflow notification emails.
-   *
-   * Refuses to send when the shared inbox prerequisites aren't met because
-   * the demo path can't recover a Reply-To without it — at that point the
-   * user must attach a real outbound email provider.
-   */
-  private async sendViaNovuDemoProvider(
-    config: ResolvedAgentConfig,
-    params: {
-      from: string;
-      to: string;
-      subject: string;
-      html: string;
-      text?: string;
-      alternatives?: Array<{ contentType: string; content: string | Buffer }>;
-      inReplyTo?: string;
-      references?: string;
-      messageId?: string;
-    },
-    integration: IntegrationEntity
-  ): Promise<{ messageId?: string }> {
-    if (!isAgentSharedInboxEnabled() || !config.credentials.emailSlugPrefix || !config.credentials.inboxRoutingKey) {
-      throw new BadRequestException(
-        'Email agent integration requires either a shared agent inbox or a custom outbound email provider. ' +
-          'Configure one in the agent email setup.'
-      );
-    }
-
-    if (config.credentials.sharedInboxDisabled) {
-      throw new BadRequestException(
-        'The Novu demo sender requires the shared inbox to be enabled. ' +
-          'Re-enable it or attach an outbound email provider.'
-      );
-    }
-
-    const limit = await this.calculateLimitNovuIntegration.execute({
-      channelType: ChannelTypeEnum.EMAIL,
-      environmentId: config.environmentId,
-      organizationId: config.organizationId,
-    });
-    if (limit && limit.count >= limit.limit) {
-      throw new BadRequestException(
-        `Novu demo email quota exhausted for this environment (${limit.count}/${limit.limit} this month). Attach an outbound email provider (e.g. SendGrid) to remove this cap.`
-      );
-    }
-
-    if (!areNovuEmailCredentialsSet()) {
-      throw new BadRequestException(
-        'Novu demo email is not configured on this deployment. Attach an outbound email provider to send replies.'
-      );
-    }
-
-    const from = buildAgentSharedInbox(config.credentials.emailSlugPrefix, config.credentials.inboxRoutingKey);
-    const senderName = resolveAgentEmailSenderName(config);
-
-    // The Novu demo integration row's stored credentials are empty by design —
-    // the real API key lives in the deployment's env so a single Novu-managed
-    // SendGrid account fans out across every org's demo sender. We rebuild
-    // credentials on every send rather than mutating the row.
-    const demoIntegration: IntegrationEntity = {
-      ...integration,
-      credentials: {
-        apiKey: process.env.NOVU_EMAIL_INTEGRATION_API_KEY,
-        from,
-        senderName,
-        ipPoolName: 'Demo',
-      },
-    };
-
-    const mailFactory = new MailFactory();
-    const handler = mailFactory.getHandler(demoIntegration, from);
-
-    const mailOptions: IEmailOptions = {
-      to: [params.to],
-      subject: params.subject,
-      html: params.html,
-      text: params.text,
-      alternatives: params.alternatives,
-      from,
-      senderName,
-      headers: {
-        ...(params.messageId ? { 'Message-ID': wrapMsgId(params.messageId) } : {}),
-        ...(params.inReplyTo ? { 'In-Reply-To': wrapMsgId(params.inReplyTo) } : {}),
-        ...(params.references
-          ? { References: params.references.split(/\s+/).filter(Boolean).map(wrapMsgId).join(' ') }
-          : {}),
-      },
-    };
-
-    const result = await handler.send(mailOptions).catch(toDeliveryError);
-
-    const messageIdForReturn = result?.id || params.messageId || '';
-
-    try {
-      await this.messageRepository.create({
-        _environmentId: config.environmentId,
-        _organizationId: config.organizationId,
-        channel: ChannelTypeEnum.EMAIL,
-        providerId: EmailProviderIdEnum.Novu,
-        email: params.to,
-        subject: params.subject,
-        transactionId: messageIdForReturn || randomUUID(),
-        payload: {
-          agentId: config.agentId,
-          html: params.html,
-          text: params.text,
-        },
-        tags: ['agent-demo-reply'],
-      });
-    } catch (err) {
-      this.logger.warn(
-        { err, environmentId: config.environmentId, agentId: config.agentId },
-        'Failed to persist Novu demo email message for quota accounting'
-      );
-      captureAgentWarning(err, {
-        component: 'chat-sdk',
-        operation: 'persist-demo-email-quota',
-        agentId: config.agentId,
-        extra: { environmentId: config.environmentId },
-      });
-    }
-
-    return { messageId: messageIdForReturn };
-  }
-
   private async createChatInstance(
     instanceKey: string,
     agentId: string,
@@ -1446,7 +730,7 @@ export class ChatSdkService implements OnModuleDestroy {
           email: createNovuEmailAdapter({
             senderName: resolveAgentEmailSenderName(config),
             signingSecret: credentials.secretKey,
-            sendEmail: this.buildSendEmailCallback(config, outboundIntegrationId),
+            sendEmail: this.agentEmailSender.buildSendEmailCallback(config, outboundIntegrationId),
             actionUrlBuilder: async ({ threadId, messageId, actionId, value, label, style }) => {
               const userIdentifier = extractRecipientFromThreadId(threadId);
               const { url } = await this.actionTokenService.signActionToken({
